lesson03/python_client/pyproject.toml
[project]
name = "python-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "grpcio-tools>=1.68.1",
    "grpcio>=1.68.1",
    "protobuf>=5.29.2",
]
